Move db to its own file
This means other files can import the database without the possibility
of a circular dependence with app/__init__.py (at least where the
database is concerned).
